+++ dbms/src/Functions/concat.cpp
@@ -1,0 +1,151 @@
+#include <DataTypes/DataTypeString.h>
+#include <Columns/ColumnString.h>
+#include <Functions/FunctionFactory.h>
+#include <Functions/FunctionHelpers.h>
+#include <Functions/IFunction.h>
+#include <Functions/GatherUtils/GatherUtils.h>
+#include <IO/WriteHelpers.h>
+#include <ext/range.h>
+
+
+namespace DB
+{
+
+namespace ErrorCodes
+{
+    extern const int ILLEGAL_TYPE_OF_ARGUMENT;
+    extern const int NUMBER_OF_ARGUMENTS_DOESNT_MATCH;
+}
+
+using namespace GatherUtils;
+
+
+template <typename Name, bool is_injective>
+class ConcatImpl : public IFunction
+{
+public:
+    static constexpr auto name = Name::name;
+    ConcatImpl(const Context & context) : context(context) {}
+    static FunctionPtr create(const Context & context)
+    {
+        return std::make_shared<ConcatImpl>(context);
+    }
+
+    String getName() const override
+    {
+        return name;
+    }
+
+    bool isVariadic() const override
+    {
+        return true;
+    }
+
+    size_t getNumberOfArguments() const override
+    {
+        return 0;
+    }
+
+    bool isInjective(const Block &) override
+    {
+        return is_injective;
+    }
+
+    bool useDefaultImplementationForConstants() const override { return true; }
+
+    DataTypePtr getReturnTypeImpl(const DataTypes & arguments) const override
+    {
+        if (!is_injective && !arguments.empty() && isArray(arguments[0]))
+            return FunctionArrayConcat(context).getReturnTypeImpl(arguments);
+
+        if (arguments.size() < 2)
+            throw Exception("Number of arguments for function " + getName() + " doesn't match: passed " + toString(arguments.size())
+                + ", should be at least 2.",
+                ErrorCodes::NUMBER_OF_ARGUMENTS_DOESNT_MATCH);
+
+        for (const auto arg_idx : ext::range(0, arguments.size()))
+        {
+            const auto arg = arguments[arg_idx].get();
+            if (!isStringOrFixedString(arg))
+                throw Exception{"Illegal type " + arg->getName() + " of argument " + std::to_string(arg_idx + 1) + " of function " + getName(),
+                    ErrorCodes::ILLEGAL_TYPE_OF_ARGUMENT};
+        }
+
+        return std::make_shared<DataTypeString>();
+    }
+
+    void executeImpl(Block & block, const ColumnNumbers & arguments, size_t result, size_t input_rows_count) override
+    {
+        if (!is_injective && !arguments.empty() && isArray(block.getByPosition(arguments[0]).type))
+            return FunctionArrayConcat(context).executeImpl(block, arguments, result, input_rows_count);
+
+        if (arguments.size() == 2)
+            executeBinary(block, arguments, result, input_rows_count);
+        else
+            executeNAry(block, arguments, result, input_rows_count);
+    }
+
+private:
+    const Context & context;
+
+    void executeBinary(Block & block, const ColumnNumbers & arguments, const size_t result, size_t input_rows_count)
+    {
+        const IColumn * c0 = block.getByPosition(arguments[0]).column.get();
+        const IColumn * c1 = block.getByPosition(arguments[1]).column.get();
+
+        const ColumnString * c0_string = checkAndGetColumn<ColumnString>(c0);
+        const ColumnString * c1_string = checkAndGetColumn<ColumnString>(c1);
+        const ColumnConst * c0_const_string = checkAndGetColumnConst<ColumnString>(c0);
+        const ColumnConst * c1_const_string = checkAndGetColumnConst<ColumnString>(c1);
+
+        auto c_res = ColumnString::create();
+
+        if (c0_string && c1_string)
+            concat(StringSource(*c0_string), StringSource(*c1_string), StringSink(*c_res, c0->size()));
+        else if (c0_string && c1_const_string)
+            concat(StringSource(*c0_string), ConstSource<StringSource>(*c1_const_string), StringSink(*c_res, c0->size()));
+        else if (c0_const_string && c1_string)
+            concat(ConstSource<StringSource>(*c0_const_string), StringSource(*c1_string), StringSink(*c_res, c0->size()));
+        else
+        {
+            /// Fallback: use generic implementation for not very important cases.
+            executeNAry(block, arguments, result, input_rows_count);
+            return;
+        }
+
+        block.getByPosition(result).column = std::move(c_res);
+    }
+
+    void executeNAry(Block & block, const ColumnNumbers & arguments, const size_t result, size_t input_rows_count)
+    {
+        size_t num_sources = arguments.size();
+        StringSources sources(num_sources);
+
+        for (size_t i = 0; i < num_sources; ++i)
+            sources[i] = createDynamicStringSource(*block.getByPosition(arguments[i]).column);
+
+        auto c_res = ColumnString::create();
+        concat(sources, StringSink(*c_res, input_rows_count));
+        block.getByPosition(result).column = std::move(c_res);
+    }
+};
+
+struct NameConcat
+{
+    static constexpr auto name = "concat";
+};
+struct NameConcatAssumeInjective
+{
+    static constexpr auto name = "concatAssumeInjective";
+};
+
+using FunctionConcat = ConcatImpl<NameConcat, false>;
+using FunctionConcatAssumeInjective = ConcatImpl<NameConcatAssumeInjective, true>;
+
+void registerFunctionsConcat(FunctionFactory & factory)
+{
+    factory.registerFunction<FunctionConcat>();
+    factory.registerFunction<FunctionConcatAssumeInjective>();
+}
+
+}
