+++ dbms/src/Functions/endsWith.cpp
@@ -1,0 +1,16 @@
+#include <DataTypes/DataTypeString.h>
+#include <Functions/FunctionFactory.h>
+#include <Functions/FunctionStartsEndsWith.h>
+
+
+namespace DB
+{
+
+using FunctionEndsWith = FunctionStartsEndsWith<NameEndsWith>;
+
+void registerFunctionEndsWith(FunctionFactory & factory)
+{
+    factory.registerFunction<FunctionEndsWith>();
+}
+}
+
