+++ dbms/src/Functions/lower.cpp
@@ -1,0 +1,21 @@
+#include <DataTypes/DataTypeString.h>
+#include <Functions/FunctionFactory.h>
+#include <Functions/FunctionStringToString.h>
+#include <Functions/LowerUpperImpl.h>
+
+
+namespace DB
+{
+
+struct NameLower
+{
+    static constexpr auto name = "lower";
+};
+using FunctionLower = FunctionStringToString<LowerUpperImpl<'A', 'Z'>, NameLower>;
+
+void registerFunctionLower(FunctionFactory & factory)
+{
+    factory.registerFunction<FunctionLower>();
+}
+
+}
