+++ dbms/src/Functions/startsWith.cpp
@@ -1,0 +1,16 @@
+#include <DataTypes/DataTypeString.h>
+#include <Functions/FunctionFactory.h>
+#include <Functions/FunctionStartsEndsWith.h>
+
+
+namespace DB
+{
+
+using FunctionStartsWith = FunctionStartsEndsWith<NameStartsWith>;
+
+void registerFunctionStartsWith(FunctionFactory & factory)
+{
+    factory.registerFunction<FunctionStartsWith>();
+}
+
+}
